+++ components/cdc/tests/integrations/test_cdc.rs
@@ -1,13 +1,11 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::cell::Cell;
-use std::rc::Rc;
 use std::sync::*;
 use std::time::Duration;
 
 use futures::sink::Sink;
-use futures::{Future, Stream};
-use grpcio::{ClientDuplexReceiver, ClientDuplexSender, WriteFlags};
+use futures::Future;
+use grpcio::WriteFlags;
 #[cfg(not(feature = "prost-codec"))]
 use kvproto::cdcpb::*;
 #[cfg(feature = "prost-codec")]
@@ -19,41 +17,9 @@ use kvproto::cdcpb::{
 use kvproto::kvrpcpb::*;
 use pd_client::PdClient;
 use test_raftstore::sleep_ms;
+use util::*;
 
 use cdc::Task;
-
-use super::TestSuite;
-
-#[allow(clippy::type_complexity)]
-fn new_event_feed(
-    client: &ChangeDataClient,
-) -> (
-    ClientDuplexSender<ChangeDataRequest>,
-    Rc<Cell<Option<ClientDuplexReceiver<ChangeDataEvent>>>>,
-    impl Fn(bool) -> Event_oneof_event,
-) {
-    let (req_tx, resp_rx) = client.event_feed().unwrap();
-    let event_feed_wrap = Rc::new(Cell::new(Some(resp_rx)));
-    let event_feed_wrap_clone = event_feed_wrap.clone();
-
-    let receive_event = move |keep_resolved_ts: bool| loop {
-        let event_feed = event_feed_wrap_clone.as_ref();
-        let (change_data, events) = match event_feed.replace(None).unwrap().into_future().wait() {
-            Ok(res) => res,
-            Err(e) => panic!("receive failed {:?}", e.0),
-        };
-        event_feed.set(Some(events));
-        let mut change_data = change_data.unwrap();
-        assert_eq!(change_data.events.len(), 1);
-        let change_data_event = &mut change_data.events[0];
-        let event = change_data_event.event.take().unwrap();
-        match event {
-            Event_oneof_event::ResolvedTs(_) if !keep_resolved_ts => continue,
-            other => return other,
-        }
-    };
-    (req_tx, event_feed_wrap, receive_event)
-}
 
 #[test]
 fn test_cdc_basic() {
@@ -579,68 +545,6 @@ fn test_region_split() {
     assert_eq!(req.get_region_id(), region.get_id());
     req.set_region_epoch(region.get_region_epoch().clone());
 
-    let (req_tx, resp_rx) = suite.get_cdc_client(1).event_feed().unwrap();
-    let _ = req_tx.send((req, WriteFlags::default())).wait().unwrap();
-    event_feed_wrap.replace(Some(resp_rx));
-    let event = receive_event(false);
-    match event {
-        Event_oneof_event::Entries(es) => {
-            assert!(es.entries.len() == 1, "{:?}", es);
-            let e = &es.entries[0];
-            assert_eq!(e.get_type(), EventLogType::Initialized, "{:?}", es);
-        }
-        _ => panic!("unknown event"),
-    }
-
-    event_feed_wrap.as_ref().replace(None);
-    suite.stop();
-}
-
-#[test]
-fn test_failed_pending_batch() {
-    let _guard = super::setup_fail();
-    let mut suite = TestSuite::new(3);
-
-    let incremental_scan_fp = "before_schedule_incremental_scan";
-    fail::cfg(incremental_scan_fp, "pause").unwrap();
-
-    let region = suite.cluster.get_region(&[]);
-    let mut req = ChangeDataRequest::default();
-    req.region_id = region.get_id();
-    req.set_region_epoch(region.get_region_epoch().clone());
-    let (req_tx, event_feed_wrap, receive_event) = new_event_feed(suite.get_cdc_client(1));
-    let _ = req_tx
-        .send((req.clone(), WriteFlags::default()))
-        .wait()
-        .unwrap();
-    // Split region.
-    suite.cluster.must_split(&region, b"k0");
-    // Wait for receiving split cmd.
-    sleep_ms(200);
-    fail::remove(incremental_scan_fp);
-
-    let event = receive_event(false);
-    match event {
-        Event_oneof_event::Entries(es) => {
-            assert!(es.entries.len() == 1, "{:?}", es);
-            let e = &es.entries[0];
-            assert_eq!(e.get_type(), EventLogType::Initialized, "{:?}", es);
-        }
-        _ => panic!("unknown event"),
-    }
-    let event = receive_event(false);
-    match event {
-        Event_oneof_event::Error(err) => {
-            assert!(err.has_epoch_not_match(), "{:?}", err);
-        }
-        _ => panic!("unknown event"),
-    }
-
-    // Try to subscribe region again.
-    let region = suite.cluster.get_region(b"k0");
-    // Ensure it is old region.
-    assert_eq!(req.get_region_id(), region.get_id());
-    req.set_region_epoch(region.get_region_epoch().clone());
     let (req_tx, resp_rx) = suite.get_cdc_client(1).event_feed().unwrap();
     let _ = req_tx.send((req, WriteFlags::default())).wait().unwrap();
     event_feed_wrap.replace(Some(resp_rx));
