+++ components/cdc/tests/util/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "util"
+version = "0.0.1"
+edition = "2018"
+publish = false
+
+[features]
+protobuf-codec = [
+  "grpcio/protobuf-codec",
+  "kvproto/protobuf-codec",
+  "raftstore/protobuf-codec",
+  "tikv/protobuf-codec",
+  "txn_types/protobuf-codec",
+]
+prost-codec = [
+  "grpcio/prost-codec",
+  "kvproto/prost-codec",
+  "raftstore/prost-codec",
+  "tikv/prost-codec",
+  "txn_types/prost-codec",
+]
+
+[dependencies]
+cdc = {path = "../../" }
+futures = "0.1"
+grpcio = { version = "0.5.0-alpha.3", default-features = false, features = ["openssl-vendored"] }
+kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
+raftstore = { path = "../../../raftstore" }
+slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
+slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
+tikv = { path = "../../../../", default-features = false }
+tikv_util = { path = "../../../tikv_util" }
+txn_types = { path = "../../../txn_types" }
+test_raftstore = { path = "../../../test_raftstore" }
+test_util = { path = "../../../test_util" }
