+++ tests/failpoints/cases/mod.rs
@@ -1,7 +1,6 @@
 // Copyright 2017 TiKV Project Authors. Licensed under Apache-2.0.
 
 mod test_bootstrap;
-mod test_cdc;
 mod test_conf_change;
 mod test_coprocessor;
 mod test_gc_worker;
